scaffolding/python3/gRPC_protobuf/requirements.txt
protobuf
grpcio-tools
futures

